Create README.md
Hi everyone. This is Mikey here; I'm currently studying to become a security analyst or hopefully a pen tester. Cyber security is my bag and I really am putting lots of work into this. If , you'd like to collaborate just lmk. Not an expert, but I am learning at a very fast rate. Thanks in advance for all the cool tools and resources.